+++ CHANGELOG.md
@@ -1808,3 +1808,4 @@
 - [2026-06-06 11:29] test(analytics): implement widget integration tests for analytics
 - [2026-06-06 12:23] feat(network): implement shimmer loading skeleton in network
 - [2026-06-06 13:46] refactor(test): extract reusable atomic widgets from test
+- [2026-06-06 15:09] feat(network): add bi-directional event bus handler for network